worker/requirements.txt
redis==5.0.7
pymongo==4.8.0
python-dotenv==1.0.1
pytest==8.3.2
pytest-cov==5.0.0
